requirements.txt
annotated-types==0.7.0
anyio==4.8.0
atproto==0.0.57
blurhash==1.1.4
certifi==2022.9.24
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.8
cryptography==43.0.3
decorator==5.1.1
dnspython==2.7.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.4
libipld==3.0.0
Mastodon.py==1.8.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2022.6
requests==2.28.1
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==1.26.12
websockets==13.1
